pyproject.toml
[project]
name = "JudgementCall"
version = "0.1.0"
description = "A civic education web app for judges and judicial elections."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cssselect>=1.4.0",
    "django>=5.2,<6",
    "django-debug-toolbar>=5.1.0",
    "django-environ>=0.12.0,<1",
    "django-localflavor>=5.0",
    "django-structlog>=9.1.0,<10",
    "django-typer>=3.1.0,<4",
    "faker>=40.15.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "lxml>=6.0.4",
    "pandas>=3.0.2",
    "pre-commit>=4.2.0",
    "pytest-django>=4.11.1,<5",
    "requests>=2.33.1",
    "rich>=14.0.0",
    "ruff>=0.11.4",
    "us>=3.2.0",
    "whitenoise>=6.9.0,<7",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors - handles basic Python style
  "W", # pycodestyle warnings - handles whitespace issues
  "F", # pyflakes - handles unused imports and variables
  "N", # pep8-naming - handles naming conventions
]
ignore = [
  "F401", # allow unused imports (useful for assignments)
  "N818",
]

# when linting, ignore the lines limited to 100 chars rule with migrations
# because they're always super long and throw errors in the pre-commit
[tool.ruff.lint.per-file-ignores]
"*/migrations/*" = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]

